backend/requirements.txt
django-filter==2.0.0
markdown==3.0.1
djangorestframework==3.9.1
django==2.1.5
mixer==6.1.3
graphene-django==2.2.0
djangorestframework-jwt==1.11.0
django-cors-headers==2.4.0
uWSGI==2.0.17
pytest==4.1.1
pytest-django==3.4.5
pytest-cov==2.6.1
python-status==1.0.1
